Refactor template info calculations
