+++ packages/edit-site/src/components/header/document-actions/index.js
@@ -20,6 +20,7 @@ import { useRef } from '@wordpress/element';
  * Internal dependencies
  */
 import TemplateDetails from '../../template-details';
+import { getTemplateInfo } from '../../../utils';
 
 function getBlockDisplayText( block ) {
 	return block
@@ -50,7 +51,7 @@ function useSecondaryText() {
 }
 
 export default function DocumentActions( { template } ) {
-	const documentTitle = template?.slug;
+	const { title: documentTitle } = getTemplateInfo( template );
 	const { label, isActive } = useSecondaryText();
 
 	// Title is active when there is no secondary item, or when the secondary
