+++ packages/edit-site/src/utils/index.js
@@ -1,1 +1,2 @@
 export { default as findTemplate } from './find-template';
+export { default as getTemplateInfo } from './get-template-info';
